Vr 28

+++ 2bim/src/ValidaCPF.java
@@ -25,7 +25,7 @@ public class ValidaCPF {
 // converte o i-esimo caractere do CPF em um numero:
 // por exemplo, transforma o caractere '0' no inteiro 0         
 // (48 eh a posicao de '0' na tabela ASCII)         
-        num = (int)(CPF.charAt(i) - 48); 
+        num = CPF.charAt(i) - 48; 
         sm = sm + (num * peso);
         peso = peso - 1;
       }
@@ -39,7 +39,7 @@ public class ValidaCPF {
       sm = 0;
       peso = 11;
       for(i=0; i<10; i++) {
-        num = (int)(CPF.charAt(i) - 48);
+        num = CPF.charAt(i) - 48;
         sm = sm + (num * peso);
         peso = peso - 1;
       }
